chỉnh csdl và các folder model, controller, service, .... ( Trân + A. Khoa + Khải + Thái làm chung)

+++ QuizClear/src/main/java/com/uth/quizclear/repository/ExamRepository.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 import com.uth.quizclear.model.Exam;
-import com.uth.quizclear.model.ExamStatus;
+import com.uth.quizclear.model.Exam.ExamStatus;
 
 @Repository
-public interface ExamRepository extends JpaRepository<Exam, Integer> {
+public interface ExamRepository extends JpaRepository<Exam, Long> { // Đổi từ Integer thành Long
     List<Exam> findByStatus(ExamStatus status);
 }
