+++ QuizClear/src/main/java/com/uth/quizclear/repository/ExamReviewRepository.java
@@ -1,11 +1,11 @@
 package com.uth.quizclear.repository;
 
-import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import com.uth.quizclear.model.ExamReview;
 
 @Repository
-public interface ExamReviewRepository extends JpaRepository<ExamReview, Integer> {
-    
+public interface ExamReviewRepository extends JpaRepository<ExamReview, Long> {
+    // Có thể thêm các method custom nếu cần
 }
